+++ src/utils/minifiy.spec.ts
@@ -220,6 +220,8 @@ describe("minify", () => {
     [`solid #00f 0.63em;`, "solid#00f.63em"],
     [`solid #f00a 0.63em;`, "solid#f00a.63em"],
     [`solid transparent 0.63em;`, "solid#0000.63em"],
+    [`transparent 0.0 0.1ch`, "#0000.0.1ch"],
+    [`transparent 0 0.1ch`, "#0000.0.1ch"],
     [
       `<style>
 * {